updated components with TypeScript
