+++ package.json
@@ -10,6 +10,11 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@types/react": "^18.2.23",
+    "@types/react-dom": "^18.2.7",
+    "@types/react-redux": "^7.1.26",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/styled-components": "^5.1.28",
     "bootstrap": "^5.3.1",
     "clsx": "^2.0.0",
     "react": "^18.2.0",
@@ -47,5 +52,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "typescript": "^5.2.2"
   }
 }
