+++ src/components/Cinema.tsx
@@ -5,34 +5,39 @@ import { Row, Col, Button, Container } from 'react-bootstrap';
 
 import { CinemaScreen, HomeButton, StyledContainer } from '../styled/Cinema';
 
-const Cinema = () => {
+interface CinemaProps {
+  date: string;
+  selectedSession: string;
+}
+
+const Cinema: React.FC<CinemaProps> = () => {
   const { date, selectedSession } = useParams();
   const navigate = useNavigate();
   const dispatch = useDispatch();
-  const availableSeats = useSelector((state) => state.cinema.availableSeats) || [];
-  const selectedSeats = useSelector((state) => state.cinema.selectedSeats);
+  const availableSeats:string[] = useSelector((state: any) => state.cinema.availableSeats) || [];
+  const selectedSeats:string[] = useSelector((state: any) => state.cinema.selectedSeats);
 
-  const [isAlertActive, setIsAlertActive] = useState(false);
+  const [isAlertActive, setIsAlertActive] = useState<boolean>(false);
 
   useEffect(() => {
     dispatch({ type: 'FETCH_AVAILABLE_SEATS' });
   }, [dispatch]);
 
-  const isSeatBooked = (seat) => selectedSeats.includes(seat);
+  const isSeatBooked = (seat: string):boolean => selectedSeats.includes(seat);
 
 
-  const handleSeatSelect = (seat) => {
+  const handleSeatSelect = (seat:string):void => {
     dispatch({ type: 'TOGGLE_SELECTED_SEAT', payload: seat });
   };
 
-  const handleBookSeats = () => {
+  const handleBookSeats = ():void => {
     alert(`Booked seats: ${selectedSeats.join(', ')} for ${selectedSession}`);
     setIsAlertActive(true);
   };
 
-  const areSeatsSelected = selectedSeats.length > 0;
+  const areSeatsSelected:boolean = selectedSeats.length > 0;
 
-  const handleGoMainPage = () => {
+  const handleGoMainPage = (): void => {
     navigate('/');
   };
 
@@ -72,4 +77,6 @@ const Cinema = () => {
   );
 };
 
+
 export default Cinema;
+
