+++ src/components/index.js
@@ -1,7 +1,7 @@
 import Cinema from './Cinema';
 import FilmDescription from './FilmDescription';
-import MainPage from './MainPage';
 import MovieSessions from './MovieSessions';
 import SessionList from './SessionList';
+import MainPage from './MainPage';
 
 export { Cinema, FilmDescription, MainPage, MovieSessions, SessionList };
