+++ src/redux/store.js
@@ -31,3 +31,5 @@ sagaMiddleware.run(filmDescriptionSaga);
 sagaMiddleware.run(cinemaSaga);
 
 export default store;
+
+
