+++ src/styled/global.d.ts
@@ -1,0 +1,14 @@
+// global.d.ts
+declare module 'global' {
+    export const colors: {
+      main: string;
+      primary: string;
+      secondary: string;
+    };
+  
+    export const fontSize: {
+      medium: string;
+      large: string;
+    };
+  }
+
